Attachments fixes

+++ shared/component/main/src/commonMain/kotlin/com/sedsoftware/tackle/main/integration/editor/EditorTabComponentApi.kt
@@ -18,7 +18,7 @@ internal class EditorTabComponentApi(
 
     override suspend fun sendFile(
         file: PlatformFileWrapper,
-        onUpload: (Int) -> Unit,
+        onUpload: (Float) -> Unit,
         thumbnail: PlatformFileWrapper?,
         description: String?,
         focus: String?,
