+++ shared/component/tab/editor/src/commonTest/kotlin/com/sedsoftware/tackle/editor/stubs/EditorTabComponentApiStub.kt
@@ -37,7 +37,7 @@ class EditorTabComponentApiStub : StubWithException(), EditorTabComponentGateway
 
     override suspend fun sendFile(
         file: PlatformFileWrapper,
-        onUpload: (Int) -> Unit,
+        onUpload: (Float) -> Unit,
         thumbnail: PlatformFileWrapper?,
         description: String?,
         focus: String?,
